singly linked-list/distro-systems
